# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HotelMate API
  version: 1.0.0
  description: Production-ready tenant-aware hotel operations, revenue reporting, approved-knowledge conversations, security audit trails, and live events. Every HTTP response includes an X-Request-ID correlation header.
servers:
  - url: /api/v1
security: []
paths:
...
  /public/hotels/{slug}/content:
    get:
      summary: Browse active public hotel content
      parameters:
        - $ref: '#/components/parameters/HotelSlug'
      responses:
        '200':
          description: Active facilities, current promotions, restaurants, and available menu items
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HotelContent' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /staff/content:
    get:
      summary: Read all tenant hotel content, including inactive records
      security: [{ staffBearer: [] }]
      responses:
        '200':
          description: Tenant-scoped editable content
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HotelContent' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /staff/facilities:
    post:
      summary: Create a hotel facility
      description: Restricted to hotel administrators and operations managers.
      security: [{ staffBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FacilityInput' }
      responses:
        '201':
          description: Facility created
          content:
            application/json:
              schema:
                type: object
                properties:
                  facility: { $ref: '#/components/schemas/Facility' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /staff/facilities/{id}:
    patch:
      summary: Replace editable facility fields
      description: Restricted to hotel administrators and operations managers.
      security: [{ staffBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FacilityInput' }
      responses:
        '200':
          description: Facility updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  facility: { $ref: '#/components/schemas/Facility' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /staff/promotions:
    post:
      summary: Create a promotion
      description: Restricted to hotel administrators and operations managers.
      security: [{ staffBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PromotionInput' }
      responses:
        '201':
          description: Promotion created
          content:
            application/json:
              schema:
                type: object
                properties:
                  promotion: { $ref: '#/components/schemas/Promotion' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /staff/promotions/{id}:
    patch:
      summary: Replace editable promotion fields
      description: Restricted to hotel administrators and operations managers.
      security: [{ staffBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PromotionInput' }
      responses:
        '200':
          description: Promotion updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  promotion: { $ref: '#/components/schemas/Promotion' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /staff/restaurants:
    post:
      summary: Create a restaurant
      description: Restricted to hotel administrators and operations managers.
      security: [{ staffBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RestaurantInput' }
      responses:
        '201':
          description: Restaurant created
          content:
            application/json:
              schema:
                type: object
                properties:
                  restaurant: { $ref: '#/components/schemas/Restaurant' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /staff/restaurants/{id}:
    patch:
      summary: Replace editable restaurant fields
      description: Restricted to hotel administrators and operations managers.
      security: [{ staffBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RestaurantInput' }
      responses:
        '200':
          description: Restaurant updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  restaurant: { $ref: '#/components/schemas/Restaurant' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /staff/restaurants/{id}/menu-items:
    post:
      summary: Add an item to a tenant restaurant menu
      description: Restricted to hotel administrators and operations managers.
      security: [{ staffBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MenuItemInput' }
      responses:
        '201':
          description: Menu item created
          content:
            application/json:
              schema:
                type: object
                properties:
                  menuItem: { $ref: '#/components/schemas/MenuItem' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /staff/menu-items/{id}:
    patch:
      summary: Replace editable menu item fields
      description: Restricted to hotel administrators and operations managers.
      security: [{ staffBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/ResourceID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MenuItemInput' }
      responses:
        '200':
          description: Menu item updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  menuItem: { $ref: '#/components/schemas/MenuItem' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
components:
  parameters:
    HotelSlug:
      name: slug
      in: path
      required: true
      schema: { type: string }
    ResourceID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  securitySchemes:
    staffBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT with the hotelmate:staff audience.
    guestBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT with the hotelmate:guest audience and an active stay.
  schemas:
    Hotel:
      type: object
      required: [id, name, slug, logoUrl, primaryColor, timezone]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        logoUrl: { type: string }
        primaryColor: { type: string }
        timezone: { type: string }
...
    Facility:
      type: object
      required: [id, name, description, icon, hours, sortOrder, isActive]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        icon: { type: string }
        hours: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    FacilityInput:
      type: object
      required: [name, description, icon, hours, sortOrder, isActive]
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        description: { type: string, maxLength: 500 }
        icon: { type: string, minLength: 2, maxLength: 32 }
        hours: { type: string, maxLength: 120 }
        sortOrder: { type: integer, minimum: 0, maximum: 10000 }
        isActive: { type: boolean }
    Promotion:
      type: object
      required: [id, title, description, discountPct, badgeText, startsAt, endsAt, isActive]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        discountPct: { type: number, minimum: 0, maximum: 100 }
        badgeText: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        isActive: { type: boolean }
    PromotionInput:
      type: object
      required: [title, description, discountPct, badgeText, startsAt, endsAt, isActive]
      properties:
        title: { type: string, minLength: 2, maxLength: 140 }
        description: { type: string, maxLength: 700 }
        discountPct: { type: number, minimum: 0, maximum: 100 }
        badgeText: { type: string, maxLength: 32 }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        isActive: { type: boolean }
    MenuItem:
      type: object
      required: [id, restaurantId, name, description, priceCents, currency, sortOrder, isAvailable]
      properties:
        id: { type: string, format: uuid }
        restaurantId: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        priceCents: { type: integer, format: int64, minimum: 0, description: Integer IRR amount; compatibility field name retained. }
        currency: { type: string, minLength: 3, maxLength: 3 }
        sortOrder: { type: integer }
        isAvailable: { type: boolean }
    MenuItemInput:
      type: object
      required: [name, description, priceCents, currency, sortOrder, isAvailable]
      properties:
        name: { type: string, minLength: 2, maxLength: 140 }
        description: { type: string, maxLength: 500 }
        priceCents: { type: integer, format: int64, minimum: 0, maximum: 100000000000 }
        currency: { type: string, minLength: 3, maxLength: 3 }
        sortOrder: { type: integer, minimum: 0, maximum: 10000 }
        isAvailable: { type: boolean }
    Restaurant:
      type: object
      required: [id, name, description, hours, sortOrder, isActive, menuItems]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        hours: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
        menuItems: { type: array, items: { $ref: '#/components/schemas/MenuItem' } }
    RestaurantInput:
      type: object
      required: [name, description, hours, sortOrder, isActive]
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        description: { type: string, maxLength: 500 }
        hours: { type: string, maxLength: 120 }
        sortOrder: { type: integer, minimum: 0, maximum: 10000 }
        isActive: { type: boolean }
    HotelContent:
      type: object
      required: [hotel, facilities, promotions, restaurants]
      properties:
        hotel: { $ref: '#/components/schemas/Hotel' }
        facilities: { type: array, items: { $ref: '#/components/schemas/Facility' } }
        promotions: { type: array, items: { $ref: '#/components/schemas/Promotion' } }
        restaurants: { type: array, items: { $ref: '#/components/schemas/Restaurant' } }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
  responses:
...
    Forbidden:
      description: Role is not permitted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationFailed:
      description: Input validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...